Added flag bot-obey-game-rule
Added a flag to control how bots work.
In the past, bots actually simulated fake key press.
And they followed exactly same rule as the players did.

But, it resulted missing notes that were impossible to hit.
So I had to change the bot to only look like it was playing the game.

However, I do miss the ability to use the bot sort of as a test.
So I added an option to revert to how it worked.